requirements.txt
openai==1.12.0
openai-agents==0.1.5
alpaca-py==0.10.0
pandas==2.1.1
yfinance==0.2.28
matplotlib==3.8.0
numpy==1.26.0
tabulate==0.9.0
python-dateutil==2.8.2
colorama==0.4.6
argparse==1.4.0
pydantic==2.5.2
litellm==1.17.6
python-dotenv==1.0.0 